Uploaded files to add
These files are as the system was in 2009.
The java program is monolithic in nature.
It compares strings and numbers from tabular data by reading rows from a cursor created on an external SQL batch file and looking up the corresponding row from another tabular data source.  This is inefficient if the lookup key isn't indexed.  It also doesn't match the performance of a hash-key table lookup, but it is good for limited numbers of rows.